gcs pdao refactored

+++ src/main/java/bio/terra/pdao/gcs/GcsPdao.java
@@ -1,14 +1,19 @@
 package bio.terra.pdao.gcs;
 
+import bio.terra.filesystem.FireStoreFileDao;
 import bio.terra.metadata.FSFile;
 import bio.terra.metadata.FSFileInfo;
 import bio.terra.metadata.Dataset;
+import bio.terra.metadata.FSObjectBase;
+import bio.terra.metadata.FSObjectType;
+import bio.terra.metadata.FileDataBucket;
 import bio.terra.model.FileLoadModel;
 import bio.terra.pdao.exception.PdaoException;
 import bio.terra.pdao.exception.PdaoFileCopyException;
 import bio.terra.pdao.exception.PdaoInvalidUriException;
 import bio.terra.pdao.exception.PdaoSourceFileNotFoundException;
-import com.google.api.gax.paging.Page;
+import bio.terra.resourcemanagement.metadata.google.GoogleProjectResource;
+import bio.terra.service.dataproject.DataProjectService;
 import com.google.cloud.storage.Acl;
 import com.google.cloud.storage.Blob;
 import com.google.cloud.storage.BlobId;
@@ -24,15 +29,30 @@ import java.net.URI;
 import java.net.URISyntaxException;
 import java.time.Instant;
 import java.util.List;
+import java.util.UUID;
 
 @Component
 @Profile("google")
 public class GcsPdao {
-    private GcsProjectFactory gcsProjectFactory;
+    private final GcsProjectFactory gcsProjectFactory;
+    private final DataProjectService dataProjectService;
+    private final FireStoreFileDao fileDao;
 
     @Autowired
-    public GcsPdao(GcsProjectFactory gcsProjectFactory) {
+    public GcsPdao(
+        GcsProjectFactory gcsProjectFactory,
+        DataProjectService dataProjectService,
+        FireStoreFileDao fileDao) {
         this.gcsProjectFactory = gcsProjectFactory;
+        this.dataProjectService = dataProjectService;
+        this.fileDao = fileDao;
+    }
+
+    private Storage storageForBucket(FileDataBucket fileDataBucket) {
+        GoogleProjectResource projectResource = fileDataBucket.getProjectResource()
+            .orElseThrow(IllegalStateException::new);
+        GcsProject gcsProject = gcsProjectFactory.get(projectResource.getGoogleProjectId());
+        return gcsProject.getStorage();
     }
 
     // We return the incoming FSObject with the blob information filled in
@@ -67,6 +87,8 @@ public class GcsPdao {
         // Figure out where the destination bucket is. Similar to how there is a ProjectIdSelector there will be need
         // to be a BucketSelector that picks a (Project, Bucket) pair for the file. The Resource Manager will then need
         // to potentially create the project and/or the bucket inside of that project depending on what already exists.
+        FileDataBucket bucketForFile = dataProjectService.getBucketForFile(fsObject);
+        Storage storage = storageForBucket(bucketForFile);
 
         Blob sourceBlob = storage.get(BlobId.of(sourceBucket, sourcePath));
         if (sourceBlob == null) {
@@ -84,7 +106,7 @@ public class GcsPdao {
             // I have been seeing timeouts and I think they are due to particularly large files,
             // so I changed exported the timeouts to application.properties to allow for tuning
             // and I am changing this to copy chunks.
-            CopyWriter writer = sourceBlob.copyTo(BlobId.of(gcsConfiguration.getBucket(), targetPath));
+            CopyWriter writer = sourceBlob.copyTo(BlobId.of(bucketForFile.getName(), targetPath));
             while (!writer.isDone()) {
                 writer.copyChunk();
             }
@@ -105,7 +127,7 @@ public class GcsPdao {
             Instant createTime = Instant.ofEpochMilli(targetBlob.getCreateTime());
 
             URI gspath = new URI("gs",
-                gcsConfiguration.getBucket(),
+                bucketForFile.getName(),
                 "/" + targetPath,
                 null,
                 null);
@@ -130,11 +152,12 @@ public class GcsPdao {
         }
     }
 
-    public boolean deleteFile(Dataset dataset, String fileId) {
-        String dataBucket = gcsConfiguration.getBucket();
-        String filePath = dataset.getId().toString() + "/" + fileId;
-
-        Blob blob = storage.get(BlobId.of(dataBucket, filePath));
+    public boolean deleteFile(FSFile fsFile) {
+        FileDataBucket bucketForFile = dataProjectService.getBucketForFile(fsFile);
+        Storage storage = storageForBucket(bucketForFile);
+        URI uri = URI.create(fsFile.getGspath());
+        String bucketPath = StringUtils.removeStart(uri.getPath(), "/");
+        Blob blob = storage.get(BlobId.of(bucketForFile.getName(), bucketPath));
         if (blob == null) {
             return false;
         }
@@ -142,15 +165,12 @@ public class GcsPdao {
     }
 
     public void deleteFilesFromDataset(Dataset dataset) {
-        String dataBucket = gcsConfiguration.getBucket();
-        String directory = dataset.getId().toString() + "/";
-        Page<Blob> blobs = storage.list(
-            dataBucket,
-            Storage.BlobListOption.currentDirectory(),
-            Storage.BlobListOption.prefix(directory));
-        for (Blob blob : blobs.iterateAll()) {
-            blob.delete();
-        }
+        // these files could be in multiple buckets..need to enumerate them from firestore and then iterate + delete
+        fileDao.forEachFsObjectInDataset(dataset, fsObjectBase -> {
+            if (fsObjectBase.getObjectType() != FSObjectType.DIRECTORY) {
+                deleteFile((FSFile)fsObjectBase);
+            }
+        });
     }
 
     private enum AclOp {
@@ -158,29 +178,35 @@ public class GcsPdao {
         ACL_OP_DELETE
     };
 
-    public void setAclOnFiles(String datasetId, List<String> fileIds, String readersPolicyEmail) {
-        fileAclOp(AclOp.ACL_OP_CREATE, datasetId, fileIds, readersPolicyEmail);
+    public void setAclOnFiles(Dataset dataset, List<String> fileIds, String readersPolicyEmail) {
+        fileAclOp(AclOp.ACL_OP_CREATE, dataset, fileIds, readersPolicyEmail);
     }
 
-    public void removeAclOnFiles(String datasetId, List<String> fileIds, String readersPolicyEmail) {
-        fileAclOp(AclOp.ACL_OP_DELETE, datasetId, fileIds, readersPolicyEmail);
+    public void removeAclOnFiles(Dataset dataset, List<String> fileIds, String readersPolicyEmail) {
+        fileAclOp(AclOp.ACL_OP_DELETE, dataset, fileIds, readersPolicyEmail);
     }
 
-    private void fileAclOp(AclOp op, String datasetId, List<String> fileIds, String readersPolicyEmail) {
-        String dataBucket = gcsConfiguration.getBucket();
-        String directory = datasetId + "/";
+    private void fileAclOp(AclOp op, Dataset dataset, List<String> fileIds, String readersPolicyEmail) {
         Acl.Group readerGroup = new Acl.Group(readersPolicyEmail);
         Acl acl = Acl.newBuilder(readerGroup, Acl.Role.READER).build();
+
         for (String fileId : fileIds) {
-            String blobName = directory + fileId;
-            BlobId blobId = BlobId.of(dataBucket, blobName);
-            switch (op) {
-                case ACL_OP_CREATE:
-                    storage.createAcl(blobId, acl);
-                    break;
-                case ACL_OP_DELETE:
-                    storage.deleteAcl(blobId, readerGroup);
-                    break;
+            FSObjectBase fsObjectBase = fileDao.retrieve(dataset, UUID.fromString(fileId));
+            if (fsObjectBase.getObjectType() == FSObjectType.FILE) {
+                FSFile fsFile = (FSFile)fsObjectBase;
+                FileDataBucket fileDataBucket = dataProjectService.getBucketForFile(fsFile);
+                Storage storage = storageForBucket(fileDataBucket);
+                URI gsUri = URI.create(fsFile.getGspath());
+                String bucketPath = StringUtils.removeStart(gsUri.getPath(), "/");
+                BlobId blobId = BlobId.of(fileDataBucket.getName(), bucketPath);
+                switch (op) {
+                    case ACL_OP_CREATE:
+                        storage.createAcl(blobId, acl);
+                        break;
+                    case ACL_OP_DELETE:
+                        storage.deleteAcl(blobId, readerGroup);
+                        break;
+                }
             }
         }
     }
